feat: excel function deploy
